Added Enrollment feature

+++ src/main/java/com/f1soft/sces/repository/SemesterRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SemesterRepository extends JpaRepository<Semester, Integer> {
+public interface SemesterRepository extends JpaRepository<Semester, Long> {
 
   Optional<Semester> findByLabel(String label);
 }
